requirements.txt
influxdb3-python
python-dotenv>=1.0
pika>=1.3.2
pandas